Remove usage of QStringView in ACache
For some unknown reason sometimes when QMultiHash<QStringView
is used and we try to find(query) an entry the iterator returned
is not the latest inserted, and our while loop to find the
matching ACacheValue.args will not find it and silently ignore
the return causing the caller to "leak" as it's stuck untill
the cache is cleaned.

With QString that does not happen, and eventually in Qt 6.8 we
can use the view to do lookups.